This is the 3.18.140 stable release

But Samsung doesn't strictly follow mainline with their usb driver patches,
leaving us with lots of conflicts while upstreaming. So import the entire
drivers/usb folder (except for /notifier and /manager) from A600FNXXU5CTD1,
which is linux 3.18.140 kernel. The same for security/selinux.

+++ drivers/usb/gadget/function/f_mass_storage.c
@@ -858,7 +858,7 @@ static int do_read_cd(struct fsg_common *common)
 	loff_t			file_offset, file_offset_tmp;
 	unsigned int		amount;
 	unsigned int		partial_page;
-	unsigned int		nread;
+	ssize_t			nread;
 
 	u32 nb_sectors, transfer_request;
